% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asthma}
\alias{asthma}
\title{A vector with asthma condition for 1198 samples}
\format{
asthma condition for 1198 samples
\describe{
  1 means asthma and 0 means non-asthma. Raw data are from Exposome Data Challenge 2021.
  }
}
\usage{
data(asthma)
}
\description{
A vector with asthma condition for 1198 samples
}
\keyword{datasets}
