% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expo}
\alias{expo}
\title{A dataset with 7 exposures from 1198 samples}
\format{
A data frame with 7 rows and 1198 columns:
\describe{
  7 exposures from 1198 samples. Raw data are from Exposome Data Challenge 2021.
  }
}
\usage{
data(expo)
}
\description{
A dataset with 7 exposures from 1198 samples
}
\keyword{datasets}
