% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enet.R
\name{getcf}
\alias{getcf}
\title{Get correlation cutoff by max network clusters}
\usage{
getcf(mat, cutoff = seq(0, 1, 0.02), ...)
}
\arguments{
\item{mat}{dataframe with row as feature and column as samples}

\item{cutoff}{correlation coefficient cutoff}

\item{...}{other parameters for `cor`}
}
\value{
list with cutoff
}
\description{
Get correlation cutoff by max network clusters
}
\examples{
data(meta)
cf <- getcf(t(meta))
}
