% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enet.R
\name{geteenet}
\alias{geteenet}
\title{Generate exposures-exposures network}
\usage{
geteenet(expo, cutoff = 0.9, nacf = 0.2, e = NULL, name = NULL, ...)
}
\arguments{
\item{expo}{exposures dataframe with row as exposures/peaks and column as samples}

\item{cutoff}{correlation coefficient cutoff}

\item{nacf}{NA cutoff for exposure, exposure with NA percentage larger than cutoff will be removed and others will be imputed by knn methods, default 0.2}

\item{e}{value or vector, specific exposure or peak to its export network}

\item{name}{export gml file for Cytoscape}

\item{...}{other parameters for `cor`}
}
\value{
list with index of exposures clusters, exposures clusters data and edge table
}
\description{
Generate exposures-exposures network
}
\examples{
data(expo)
ee <- geteenet(expo)
}
