% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enet.R
\name{getemnet}
\alias{getemnet}
\title{Generate metabolites-exposures network}
\usage{
getemnet(meta, expo, cutoffm = 0.9, cutoffe = 0.6, nacf = 0.2, ...)
}
\arguments{
\item{meta}{metabolites dataframe with row as metabolites/peaks and column as samples}

\item{expo}{exposures dataframe with row as exposures/peaks and column as samples}

\item{cutoffm}{correlation coefficient cutoff for metabolites-metabolites network}

\item{cutoffe}{correlation coefficient cutoff for exposures-exposures network}

\item{nacf}{NA cutoff for exposure, exposure with NA percentage larger than cutoff will be removed and others will be imputed by knn methods, default 0.2}

\item{...}{other parameters for `cor`}
}
\value{
list with table of metabolites-exposures, metabolites-exposures network clusters data and edge table
}
\description{
Generate metabolites-exposures network
}
\examples{
data(expo)
data(meta)
me <- getemnet(meta,expo)
}
