% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enet.R
\name{getgk}
\alias{getgk}
\title{Gatekeeper discovery}
\usage{
getgk(meta, expo, cutoff = 0.9, multiple = FALSE, ...)
}
\arguments{
\item{meta}{metabolites dataframe with row as metabolites/peaks and column as samples}

\item{expo}{exposures dataframe with row as exposures/peaks and column as samples}

\item{cutoff}{correlation coefficient cutoff for metabolites-metabolites network}

\item{multiple}{logical, only output gatekeeper with multiple exposures}

\item{...}{other parameters for `cor`}
}
\value{
list with table of metabolites-exposures, metabolites-exposures network clusters data and edge table
}
\description{
Gatekeeper discovery
}
\examples{
data(expo)
data(meta)
gk <- getgk(exp(meta),expo)
}
