% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enet.R
\name{getmmnet}
\alias{getmmnet}
\title{Generate metabolites-metabolites network}
\usage{
getmmnet(meta, cutoff = 0.9, m = NULL, name = NULL, ...)
}
\arguments{
\item{meta}{metabolites list with row as metabolites/peaks and column as samples}

\item{cutoff}{correlation coefficient cutoff}

\item{m}{value or vector, specific metabolite or peak to its export network}

\item{name}{export gml file for Cytoscape}

\item{...}{other parameters for `cor`}
}
\value{
list with index of metabolites clusters, metabolites clusters data and edge table
}
\description{
Generate metabolites-metabolites network
}
\examples{
data(meta)
mm <- getmmnet(meta)
}
