% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta}
\alias{meta}
\title{A dataset 177 peaks/metabolites from 1198 samples}
\format{
A data frame with 177 rows and 1198 columns:
\describe{
  177 peaks/metabolites from 1198 samples. Raw data are from Exposome Data Challenge 2021.
  }
}
\usage{
data(meta)
}
\description{
A dataset 177 peaks/metabolites from 1198 samples
}
\keyword{datasets}
